+++ src/WonderWordsSentence.php
@@ -11,6 +11,8 @@
 
 namespace NavisBorealis\WonderwordsPhp;
 
+use NavisBorealis\WonderwordsPhp\Words\Adjective;
+use NavisBorealis\WonderwordsPhp\Words\Adverb;
 use NavisBorealis\WonderwordsPhp\Words\Noun;
 use NavisBorealis\WonderwordsPhp\Words\Verb;
 
@@ -57,39 +59,60 @@ class WonderWordsSentence
     }
 
     /**
-     * Generate a bare-bone sentence in the form of
-     * `[(article)] [subject (noun)] [predicate (verb)].`
-     * Example: `A cat runs.`.
+     * Generate a bare-bone sentence.
+     * Example: `A fluffy cat quickly runs.`.
      *
-     * @param array $nounOptions options for generating the noun
-     * @param array $verbOptions options for generating the verb
+     * @param array $nounOptions      options for generating the noun
+     * @param array $verbOptions      options for generating the verb
+     * @param array $adjectiveOptions options for generating the adjective
+     * @param array $adverbOptions    options for generating the adverb
      *
      * @return string the generated sentence
      */
-    public static function bareBoneSentence(array $nounOptions = [], array $verbOptions = []): string
-    {
-        $noun = self::withArticle(Noun::randomWord($nounOptions));
+    public static function bareBoneSentence(
+        array $nounOptions = [],
+        array $verbOptions = [],
+        array $adjectiveOptions = [],
+        array $adverbOptions = []
+    ): string {
+        $adjective = Adjective::randomWord($adjectiveOptions);
+        $noun = Noun::randomWord($nounOptions);
+        $subject = self::withArticle($adjective.' '.$noun);
+
+        $adverb = Adverb::randomWord($adverbOptions);
         $verb = self::presentTense(Verb::randomWord($verbOptions));
 
-        return ucfirst($noun).' '.$verb.'.';
+        return ucfirst($subject).' '.$adverb.' '.$verb.'.';
     }
 
     /**
-     * Generate a simple sentence in the form of
-     * `[(article)] [subject (noun)] [predicate (verb)] [direct object (noun)].`
-     * Example: `A cake plays golf.`.
+     * Generate a simple sentence.
+     * Example: `A fluffy cake quickly plays golf.`.
      *
-     * @param array $nounOptions options for generating the nouns
-     * @param array $verbOptions options for generating the verb
+     * @param array $nounOptions      options for generating the nouns
+     * @param array $verbOptions      options for generating the verb
+     * @param array $adjectiveOptions options for generating the adjective
+     * @param array $adverbOptions    options for generating the adverb
      *
      * @return string the generated sentence
      */
-    public static function simpleSentence(array $nounOptions = [], array $verbOptions = []): string
-    {
-        $subject = self::withArticle(Noun::randomWord($nounOptions));
-        $verb = self::presentTense(Verb::randomWord($verbOptions));
-        $object = self::withArticle(Noun::randomWord($nounOptions));
+    public static function simpleSentence(
+        array $nounOptions = [],
+        array $verbOptions = [],
+        array $adjectiveOptions = [],
+        array $adverbOptions = []
+    ): string {
+        $adjective = Adjective::randomWord($adjectiveOptions);
+        $noun = Noun::randomWord($nounOptions);
+        $subject = self::withArticle($adjective.' '.$noun);
 
-        return ucfirst($subject).' '.$verb.' '.$object.'.';
+        $adverb = Adverb::randomWord($adverbOptions);
+        $verb = self::presentTense(Verb::randomWord($verbOptions));
+
+        $objectAdjective = Adjective::randomWord($adjectiveOptions);
+        $objectNoun = Noun::randomWord($nounOptions);
+        $object = self::withArticle($objectAdjective.' '.$objectNoun);
+
+        return ucfirst($subject).' '.$adverb.' '.$verb.' '.$object.'.';
     }
 }