+++ src/Words/Adverb.php
@@ -1,0 +1,126 @@
+<?php
+
+/*
+ * This file is part of the WonderWordsPHP package.
+ *
+ * (c) Piotr Grabski-Gradziński <piotr.gradzinski@navisborealis.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace NavisBorealis\WonderwordsPhp\Words;
+
+/**
+ * Class Adverb.
+ *
+ * Contains a dictionary of adverbs.
+ */
+class Adverb extends Words
+{
+    public static $words;
+
+    public const DEFAULT_WORDS = [
+        'accidentally',
+        'always',
+        'angrily',
+        'anxiously',
+        'awkwardly',
+        'badly',
+        'blindly',
+        'boastfully',
+        'boldly',
+        'bravely',
+        'brightly',
+        'cheerfully',
+        'coyly',
+        'crazily',
+        'defiantly',
+        'deftly',
+        'deliberately',
+        'devotedly',
+        'doubtfully',
+        'dramatically',
+        'dutifully',
+        'eagerly',
+        'elegantly',
+        'enormously',
+        'evenly',
+        'eventually',
+        'exactly',
+        'faithfully',
+        'finally',
+        'foolishly',
+        'fortunately',
+        'frequently',
+        'gleefully',
+        'gracefully',
+        'happily',
+        'hastily',
+        'honestly',
+        'hopelessly',
+        'hourly',
+        'hungrily',
+        'innocently',
+        'inquisitively',
+        'irritably',
+        'jealously',
+        'justly',
+        'kindly',
+        'lazily',
+        'loosely',
+        'madly',
+        'merrily',
+        'mysteriously',
+        'naturally',
+        'nervously',
+        'never',
+        'obediently',
+        'obnoxiously',
+        'occasionally',
+        'often',
+        'perfectly',
+        'politely',
+        'poorly',
+        'powerfully',
+        'promptly',
+        'quickly',
+        'quietly',
+        'rapidly',
+        'rarely',
+        'regularly',
+        'rudely',
+        'safely',
+        'seldom',
+        'selfishly',
+        'seriously',
+        'shakily',
+        'sharply',
+        'silently',
+        'slowly',
+        'solemnly',
+        'sometimes',
+        'speedily',
+        'sternly',
+        'strictly',
+        'successfully',
+        'suddenly',
+        'surprisingly',
+        'suspiciously',
+        'sweetly',
+        'tenderly',
+        'tensely',
+        'thoughtfully',
+        'tightly',
+        'truthfully',
+        'unexpectedly',
+        'victoriously',
+        'violently',
+        'vivaciously',
+        'warmly',
+        'weakly',
+        'wearily',
+        'wildly',
+        'yearly',
+    ];
+}