+++ src/Words/Color.php
@@ -1,0 +1,107 @@
+<?php
+
+/*
+ * This file is part of the WonderWordsPHP package.
+ *
+ * (c) Piotr Grabski-Gradziński <piotr.gradzinski@navisborealis.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace NavisBorealis\WonderwordsPhp\Words;
+
+/**
+ * Class Color.
+ *
+ * Contains a dictionary of colors.
+ */
+class Color extends Words
+{
+    public static $words;
+
+    public const DEFAULT_WORDS = [
+        'amber',
+        'amethyst',
+        'apricot',
+        'aqua',
+        'aquamarine',
+        'azure',
+        'beige',
+        'black',
+        'blue',
+        'bronze',
+        'brown',
+        'burgundy',
+        'camel',
+        'caramel',
+        'cerise',
+        'charcoal',
+        'chartreuse',
+        'chestnut',
+        'chocolate',
+        'cinnamon',
+        'cobalt',
+        'copper',
+        'coral',
+        'crimson',
+        'cyan',
+        'denim',
+        'emerald',
+        'fuchsia',
+        'gold',
+        'golden',
+        'gray',
+        'green',
+        'grey',
+        'hazel',
+        'indigo',
+        'ivory',
+        'jade',
+        'khaki',
+        'lavender',
+        'lemon',
+        'lilac',
+        'lime',
+        'magenta',
+        'maroon',
+        'mauve',
+        'mint',
+        'mustard',
+        'navy',
+        'olive',
+        'orange',
+        'orchid',
+        'peach',
+        'pear',
+        'periwinkle',
+        'pink',
+        'plum',
+        'purple',
+        'quartz',
+        'red',
+        'rose',
+        'ruby',
+        'rust',
+        'saffron',
+        'salmon',
+        'sapphire',
+        'scarlet',
+        'sepia',
+        'silver',
+        'slate',
+        'tan',
+        'tangerine',
+        'taupe',
+        'teal',
+        'tomato',
+        'topaz',
+        'turquoise',
+        'ultramarine',
+        'vanilla',
+        'violet',
+        'wheat',
+        'white',
+        'yellow',
+    ];
+}